% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.deg2rad}
\alias{.deg2rad}
\title{Degree to Radian Conversion}
\usage{
.deg2rad(x)
}
\arguments{
\item{x}{Numeric. Degree value(s).}
}
\value{
Numeric of degree value(s).
}
\description{
Converts degree value(s) to radians.
}
