% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.rad2deg}
\alias{.rad2deg}
\title{Radian to Degree Conversion}
\usage{
.rad2deg(x)
}
\arguments{
\item{x}{Numeric. Radian value(s).}
}
\value{
Numeric of degree value(s).
}
\description{
Converts radian value(s) to degrees.
}
