% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oregonElev}
\alias{oregonElev}
\title{Raster of elevation for Oregon, USA.}
\format{A raster image with 1918 x 3670 pixels.
\describe{
  \item{range}{0 -- 3332}
  \item{bounds}{-124.7039, -116.463, 41.99197, 46.29957}
  \item{resolution}{270m x 270m}
  \item{projection}{WGS84}
  \item{scalar}{1}
  ...
}}
\usage{
oregonElev
}
\description{
A raster image of Shuttle Radar Topography Mission (SRTM)
elevation over Oregon state, USA.
The elevation data were prepared (aggregated to 270m resolution)
in Google Earth Engine and downloaded on 25 September, 2019.
}
\keyword{datasets}
