% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oregonEVI}
\alias{oregonEVI}
\title{Raster of Enhanced Vegetation Index (EVI) over Oregon, USA.}
\format{A raster image with 1918 x 3670 pixels.
\describe{
  \item{range}{0 -- 10000}
  \item{bounds}{-124.7041, -116.4621, 41.99175, 46.29917}
  \item{resolution}{250m x 250m}
  \item{projection}{WGS84}
  \item{scalar}{0.0001}
  ...
}}
\usage{
oregonEVI
}
\description{
A raster image of maximum growing season (May-September)
EVI over Oregon state, USA derived from MODIS 16-day EVI data.
The growing season summary was completed in Google Earth Engine
on 30 August, 2019. Data were filtered using the general quality assurance codes
included with the MODIS data; only the highest quality (0: VI
produced with good quality) data were included.
}
\keyword{datasets}
