% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orelevation}
\alias{orelevation}
\title{SRTM elevation for a portion of southwestern Oregon, USA.}
\format{
A raster image with 371 x 371 pixels
\describe{
  \item{range}{433 -- 1390}
  \item{bounds}{-123.0001, -122.9002, 43.00015, 43.10013}
  \item{resolution}{30m x 30m (0.002694946 degrees)}
  \item{projection}{WGS84}
  \item{scalar}{1}
  ...
}
}
\usage{
orelevation
}
\description{
A raster image of Shuttle Radar Topography Mission (SRTM)
elevation for a portion of southwestern Oregon.
}
\details{
Elevation values are from the SRTM data for 2000 and are
at roughly 30m resolution. The image was created using
Google Earth Engine in October 2019.
}
\keyword{datasets}
