% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orforest}
\alias{orforest}
\title{NDVI for a portion of southwestern Oregon, USA.}
\format{A raster image with 371 x 371 pixels
\describe{
  \item{range}{0 -- 1}
  \item{bounds}{-123, -122.9, 43.0002, 43.1}
  \item{resolution}{30m x 30m (0.002694946 degrees)}
  \item{projection}{WGS84}
  \item{scalar}{1}
  ...
}}
\usage{
orforest
}
\description{
A raster image of Normalized Difference Vegetation Index (NDVI)
derived from Landsat data for a small portion of SW Oregon state.
}
\details{
NDVI values are derived from Landsat scene path 45,
row 30 summarized as the mean NDVI value between June
and August 2000 at roughly 30m resolution. Clouds
were removed from the Landsat scene before calculating
the mean. The image was created using Google Earth
Engine in August 2018.
}
\keyword{datasets}
