% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingwindow.R
\name{pad_edges}
\alias{pad_edges}
\title{Extend edges of a raster or matrix.}
\usage{
pad_edges(x, window_type = "square", size = 11, epsg_proj = 5070)
}
\arguments{
\item{x}{A raster or matrix.}

\item{window_type}{Character. Type of window, either circular or square.}

\item{size}{Numeric. Size of window, in number of pixels on each
side. For circular windows, this will be the number of pixels of the radius.}

\item{epsg_proj}{Numeric. Appropriate equal area EPSG code used to
crop raster to each circular window. Only used for circular windows.}
}
\value{
A raster with edges padded \code{size} number of pixels on each edge.
}
\description{
Extends edge values of a raster or matrix by a specified number of pixels.
}
\examples{
library(raster)

# import raster image
data(normforest)

# crop raster to much smaller area
x <- pad_edges(normforest, 'square', 11)
}
