% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rotate}
\alias{rotate}
\title{Rotates a matrix 180 degrees.}
\usage{
rotate(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A matrix rotate 180 degrees.
}
\description{
Rotates a matrix 180 degrees. Code is from https://stackoverflow.com/questions/16496210/rotate-a-matrix-in-r-by-90-degrees-clockwise.
}
