% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{sa}
\alias{sa}
\title{Calculates the Average Roughness of a Surface}
\usage{
sa(x)
}
\arguments{
\item{x}{A raster or matrix.}
}
\value{
A value of average roughness in the units of the
  original raster or matrix.
}
\description{
Finds the average roughness of a surface (Sa) as the absolute
deviation of surface heights from the mean surface height.
Height is measured as the value of a raster and may not
necessarily represent actual height.
}
\examples{
# import raster image
data(normforest)
normforest <- terra::unwrap(normforest)

# find the surface roughness
roughness <- sa(normforest)
}
