% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdq.R
\name{sdq}
\alias{sdq}
\title{Root Mean Square Slope of Surface}
\usage{
sdq(x)
}
\arguments{
\item{x}{A raster object.}
}
\value{
A numeric value representing the two-point root
  mean square slope, Sdq.
}
\description{
Calculates the root mean square slope of a raster
surface using the two-point method. This function is based
on the equations found at
https://www.ntmdt-si.ru/data/media/files/manuals/image_analisys_p9_nov12.e.pdf.
}
\examples{
# import raster image
data(normforest)

# calculate root mean square slope
Sdq <- sdq(normforest)
}
