% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localsurface.R
\name{sds}
\alias{sds}
\title{Summit Density}
\usage{
sds(x)
}
\arguments{
\item{x}{A raster or matrix.}
}
\value{
A numeric value representing the summit density.
}
\description{
Calculates the summit density of a raster or matrix. Summit density is the number of local
peaks per unit area.
}
\examples{
# import raster image
data(normforest)
normforest <- terra::unwrap(normforest)

# calculate summit density.
Sds <- sds(normforest)
}
