% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearingarea.R
\name{sk}
\alias{sk}
\title{Core Roughness Depth}
\usage{
sk(x)
}
\arguments{
\item{x}{A raster.}
}
\value{
A numeric value representing the core roughness
  depth of the image.
}
\description{
Determines the core roughness depth (Sk), the
height difference between y values of the
intersection points of the least mean square line
fit to the flattest 40\% of the bearing area curve.
See Figure 2a from Kedron et al. (2018) for more details.
}
\examples{
# import raster image
data(normforest)

# determine the core roughness depth
Sk <- sk(normforest)
}
