% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{smean}
\alias{smean}
\title{Calculates the Mean Peak Height of a Surface Raster}
\usage{
smean(x)
}
\arguments{
\item{x}{A raster.}
}
\value{
A numeric value of mean peak height.
}
\description{
Finds the mean height of positive values in the
landscape (mean peak height; Smean) for a raster
representing a surface.
}
\examples{
# import raster image
data(normforest)

# find the maximum peak height
Smean <- smean(normforest)
}
