% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{sp}
\alias{sp}
\title{Calculates the maximum peak height of a surface raster.}
\usage{
sp(x)
}
\arguments{
\item{x}{A raster.}
}
\value{
A numeric value of maximum peak height.
}
\description{
Finds the absolute value of the highest value in the
landscape (maximum peak height; Sp) for a raster
representing a surface.
}
\examples{
# import raster image
data(normforest)

# find the maximum peak height
Sp <- sp(normforest)
}
