% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{sph}
\alias{sph}
\title{Calculates the Maximum Peak Height of a Surface Raster}
\usage{
sph(x)
}
\arguments{
\item{x}{A raster.}
}
\value{
A numeric value of maximum peak height.
}
\description{
Finds the absolute value of the highest value in the
landscape (maximum peak height; Sph) for a raster
representing a surface.
}
\examples{
# import raster image
data(normforest)

# find the maximum peak height
Sph <- sph(normforest)
}
