% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{sv}
\alias{sv}
\title{Calculates the Maximum Valley Depth of a Surface Image}
\usage{
sv(x)
}
\arguments{
\item{x}{A raster or matrix.}
}
\value{
A numeric value of maximum valley depth.
}
\description{
Finds the absolute value of the lowest value in the
landscape (maximum valley depth; Sv) for a raster or matrix
representing a surface.
}
\examples{
# import raster image
data(normforest)

# find the maximum valley depth
Sv <- sv(normforest)
}
