% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sfd}
\alias{sfd}
\title{Calculate the fractal dimension of a raster.}
\usage{
sfd(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A numeric value representing the fractal dimension of
the image.
}
\description{
Calculates the 3D fractal dimension of a raster using the
triangular prism surface area method.
}
\examples{

# import raster image
data(normforest)


# convert to matrix
x <- raster::as.matrix(normforest)

# calculate the fractal dimension
Sfd <- sfd(mat)
}
