% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{sfd}
\alias{sfd}
\title{Calculates the Fractal Dimension}
\usage{
sfd(x)
}
\arguments{
\item{x}{A raster.}
}
\value{
A numeric value representing the 2D fractal dimension of
  the image.
}
\description{
Calculates the 2D fractal dimension of a raster using the
Fourier transform. This function is a modified version, adapted
to R, of the Matlab function 'fdsurfft' version 1.0.0.0
developed by Jianbo Zhang.
}
\examples{
library(raster)

# import raster image
data(normforest)

# crop raster to much smaller area
x <- crop(normforest, extent(-123, -122.99, 43, 43.01))

# calculate the fractal dimension
Sfd <- sfd(x)
}
