% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{std}
\alias{std}
\title{Texture Direction Metrics}
\usage{
std(x, create_plot = FALSE, option = c(1, 2))
}
\arguments{
\item{x}{A raster or matrix.}

\item{create_plot}{Logical. If \code{TRUE}, returns a plot of the
amplitude spectrum with lines showing directions in which
amplitude is summed for the Std and Stdi calculations.
Plotting is not possible when input is a matrix.}

\item{option}{Numeric. Code for which output metric(s) to return. 1 = Std, 2 = Stdi.}
}
\value{
A vector containing numeric values for the angle of
  dominating texture and the texture direction index.
}
\description{
Calculates the angle of dominating texture and the texture
direction index of the Fourier spectrum image calculated
from a raster image (see Kedron et al. 2018).
}
\examples{
# import raster image
data(normforest)
normforest <- terra::unwrap(normforest)

# calculate Std and Stdi
stdvals <- std(normforest)

# extract each value
Std <- stdvals[1]
Stdi <- stdvals[2]
}
