% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingwindow.R
\name{texture_image}
\alias{texture_image}
\title{Calculate Texture Metrics per Pixel}
\usage{
texture_image(x, window_type = "square", size = 11, epsg_proj = 5070,
  metric, threshold = NULL, low = NULL, high = NULL,
  parallel = TRUE, ncores = NULL)
}
\arguments{
\item{x}{A raster.}

\item{window_type}{Character. Type of window, either circular or square.}

\item{size}{Numeric. Size of window, in number of pixels on each
side for square windows (must be an odd value), or distance from
center (in meters) for circular windows.}

\item{epsg_proj}{Numeric. Appropriate equal area EPSG code used to
crop raster to each circular window.}

\item{metric}{Character. Metric to calculate for each window. Metrics
are listed below.}

\item{parallel}{Logical. Option to run the calculations in
parallel on available cores.}

\item{ncores}{Numeric. If parallel is TRUE, number of cores on which to
run the calculations. Defaults to all available, minus 1.}
}
\value{
A raster with pixel values representative of the metric
value for the window surrounding that pixel.
}
\description{
Calculates the various texture metrics over windows centered
on individual pixels. This creates a continuous surface of the
texture metric.
}
\details{
Metrics available:
\enumerate {
   \item{\code{'sa'}: average surface roughness}
   \item{\code{'sq'}: root mean square roughness}
   \item{\code{'s10z'}: ten-point height}
   \item{\code{'sdq'}: root mean square slope of surface, 2-point method}
   \item{\code{'sdq6'}: root mean square slope of surface, 7-point method}
   \item{\code{'sdr'}: surface area ratio}
   \item{\code{'sbi'}: surface bearing index}
   \item{\code{'sci'}: core fluid retention index}
   \item{\code{'ssk_adj'}: adjusted skewness}
   \item{\code{'ssk'}: skewness}
   \item{\code{'sku_exc'}: excess kurtosis}
   \item{\code{'sku'}: kurtosis}
   \item{\code{'sds'}: summit density}
   \item{\code{'sfd'}: 2d fractal dimension}
   \item{\code{'srw'}: dominant radial wavelength}
   \item{\code{'srwi'}: radial wavelength index}
   \item{\code{'shw'}: mean half wavelength}
   \item{\code{'std'}: angle of dominating texture}
   \item{\code{'stdi'}: texture direction index}
   \item{\code{'svi'}: valley fluid retention index}
   \item{\code{'str'}: texture aspect ratio}
   \item{\code{'ssc'}: mean summit curvature}
   \item{\code{'sv'}: maximum valley depth}
   \item{\code{'sph'}: maximum peak height}
   \item{\code{'sk'}: core roughness depth}
   \item{\code{'smean'}: mean peak height}
   \item{\code{'svk'}: reduced valley depth}
   \item{\code{'spk'}: reduced peak height}
   \item{\code{'scl'}: correlation length}
   \item{\code{'sdc'}: bearing area curve height interval}
}
}
\examples{
library(raster)

# import raster image
data(normforest)

# get a surface of root mean square roughness
sq_img <- texture_image(x = normforest, window = 'square',
size = 11, epsg_proj = 5070, metric = 'sq', parallel = TRUE)

# plot the result
plot(sq_img)
}
