% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingwindow.R
\name{texture_image}
\alias{texture_image}
\title{Calculate Texture Metrics per Pixel}
\usage{
texture_image(x, window_type = "square", size = 3, epsg_proj = 5070,
  metric, parallel = TRUE, ncores = NULL)
}
\arguments{
\item{x}{A raster.}

\item{window_type}{Character. Type of window, either circular or square.}

\item{size}{Numeric. Size of window, in number of pixels on each
side for square windows (must be an odd value), or distance from
center (in meters) for circular windows.}

\item{epsg_proj}{Numeric. Appropriate equal area EPSG code used to
crop raster to each circular window.}

\item{metric}{Character. Metric to calculate for each window. Metrics
are listed below.}

\item{parallel}{Logical. Option to run the calculations in
parallel on available cores.}

\item{ncores}{Numeric. If parallel is TRUE, number of cores on which to
run the calculations. Defaults to all available, minus 1.}
}
\value{
A raster with pixel values representative of the metric
value for the window surrounding that pixel.
}
\description{
Calculates the various texture metrics over windows centered
on individual pixels. This creates a continuous surface of the
texture metric.
}
\examples{
library(raster)

# import raster image
data(normforest)

# get a surface of root mean square roughness
sq_img <- texture_image(x = normforest, window = 'square',
size = 11, epsg_proj = 5070, metric = 'sq')

# plot the result
plot(sq_img)
}
