% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingwindow.R
\name{window_metric}
\alias{window_metric}
\title{Calculate Texture Metric for Single Pixel}
\usage{
window_metric(x, window_type = "square", size = 11, epsg_proj = 5070,
  rownum, colnum, metric, threshold = NULL, low = NULL, high = NULL)
}
\arguments{
\item{x}{A raster.}

\item{window_type}{Character. Type of window, either circular or square.}

\item{size}{Numeric. Size of window, in number of pixels on each
side for square windows (must be an odd value), or distance from
center (in meters) for circular windows.}

\item{epsg_proj}{Numeric. Appropriate equal area EPSG code used to
crop raster to each circular window.}

\item{rownum}{Numeric. Row number of pixel.}

\item{colnum}{Numeric. Column number of pixel.}

\item{metric}{Character. Metric to calculate for each window. Metrics
are listed below.}

\item{threshold}{Numeric. Value of autocorrelation distance (0 - 1), if
calculating \code{scl} or \code{str}.}

\item{low}{Numeric. Low value (0 - 1) if calculating \code{sdc}.}

\item{high}{Numeric. High value (0 - 1) if calculating \code{sdc}.}
}
\value{
A raster with pixel values representative of the metric
value for the window surrounding that pixel.
}
\description{
Calculates the various texture metrics over a window centered
on an individual pixel.
}
\details{
Metrics available:
\enumerate{
   \item{\code{'sa'}: average surface roughness}
   \item{\code{'sq'}: root mean square roughness}
   \item{\code{'s10z'}: ten-point height}
   \item{\code{'sdq'}: root mean square slope of surface, 2-point method}
   \item{\code{'sdq6'}: root mean square slope of surface, 7-point method}
   \item{\code{'sdr'}: surface area ratio}
   \item{\code{'sbi'}: surface bearing index}
   \item{\code{'sci'}: core fluid retention index}
   \item{\code{'ssk_adj'}: adjusted skewness}
   \item{\code{'ssk'}: skewness}
   \item{\code{'sku_exc'}: excess kurtosis}
   \item{\code{'sku'}: kurtosis}
   \item{\code{'sds'}: summit density}
   \item{\code{'sfd'}: 2d fractal dimension}
   \item{\code{'srw'}: dominant radial wavelength}
   \item{\code{'srwi'}: radial wavelength index}
   \item{\code{'shw'}: mean half wavelength}
   \item{\code{'std'}: angle of dominating texture}
   \item{\code{'stdi'}: texture direction index}
   \item{\code{'svi'}: valley fluid retention index}
   \item{\code{'str'}: texture aspect ratio}
   \item{\code{'ssc'}: mean summit curvature}
   \item{\code{'sv'}: maximum valley depth}
   \item{\code{'sph'}: maximum peak height}
   \item{\code{'sk'}: core roughness depth}
   \item{\code{'smean'}: mean peak height}
   \item{\code{'svk'}: reduced valley depth}
   \item{\code{'spk'}: reduced peak height}
   \item{\code{'scl'}: correlation length}
   \item{\code{'sdc'}: bearing area curve height interval}
}
}
\examples{
library(raster)

# import raster image
data(normforest)

# crop raster to much smaller area
x <- crop(normforest, extent(-123, -122.99, 43, 43.01))

# get a surface of root mean square roughness
sbi_img <- window_metric(x = x, window = 'circle',
size = 90, epsg_proj = 5070, rownum = 3, colnum = 3, metric = 'sbi')
}
