% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingwindow.R
\name{window_metric}
\alias{window_metric}
\title{Calculate Texture Metric for Single Pixel}
\usage{
window_metric(x, window = "square", size = 3, epsg_proj = NULL,
  rownum, colnum, metric)
}
\arguments{
\item{x}{A raster.}

\item{size}{Numeric. Size of window, in number of pixels on each
side for square windows (must be an odd value), or distance from
center (in meters) for circular windows.}

\item{epsg_proj}{Numeric. Appropriate equal area EPSG code used to
crop raster to each circular window.}

\item{rownum}{Numeric. Row number of pixel.}

\item{colnum}{Numeric. Column number of pixel.}

\item{metric}{Character. Metric to calculate for each window. Metrics
are listed below.}

\item{window_type}{Character. Type of window, either circular or square.}
}
\value{
A raster with pixel values representative of the metric
value for the window surrounding that pixel.
}
\description{
Calculates the various texture metrics over a window centered
on an individual pixel.
}
\examples{
library(raster)

# import raster image
data(normforest)

# get a surface of root mean square roughness
sq_img <- texture_image(x = normforest, window = 'circle',
size = 90, epsg_proj = 5070, metric = 'sq')

# plot the result
plot(sq_img)
}
